requirements.txt
fastapi
uvicorn
streamlit
openai
google-api-python-client
google-auth
google-auth-oauthlib
